datatables_guide/python_example/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
jinja2==3.1.2
mysql-connector-python==8.2.0
python-multipart==0.0.6
